+++ front/app/utils/moduleUtils.ts
@@ -25,6 +25,7 @@ import { IGroupDataAttributes, MembershipType } from 'services/groups';
 import { ParticipationMethod } from 'services/participationContexts';
 import {
   CellConfiguration,
+  CLErrors,
   FormikSubmitHandler,
   InsertConfigurationOptions,
   ITab,
@@ -37,6 +38,9 @@ import { IUserData } from 'services/users';
 import { MessageValue } from 'react-intl';
 import { NavItem } from 'containers/Admin/sideBar';
 import {
+  CTASignedInType,
+  CTASignedOutType,
+  CustomizedButtonConfig,
   IAppConfigurationSettings,
   TAppConfigurationSetting,
   TAppConfigurationSettingCore,
@@ -371,6 +375,24 @@ export type OutletsPropertyMap = {
     onMount: () => void;
   };
   'app.containers.Admin.settings.policies.subTitle': Record<string, any>;
+  'app.containers.Admin.settings.customize.headerSectionEnd': {
+    latestAppConfigSettings:
+      | IAppConfigurationSettings
+      | Partial<IAppConfigurationSettings>;
+    handleOnChange: (
+      settingName: TAppConfigurationSetting
+    ) => (settingKey: string, settingValue: any) => void;
+    errors: CLErrors;
+  };
+  'app.containers.LandingPage.SignedOutHeader.CTA': {
+    ctaType: CTASignedOutType;
+    customizedButtonConfig?: CustomizedButtonConfig;
+    signUpIn: (event) => void;
+  };
+  'app.containers.LandingPage.SignedInHeader.CTA': {
+    ctaType: CTASignedInType;
+    customizedButtonConfig?: CustomizedButtonConfig;
+  };
 };
 
 type Outlet<Props> = FunctionComponent<Props> | FunctionComponent<Props>[];